Automatically upload file to server in CompositeModel
