+++ doc/source/api/server_helpers.rst
@@ -8,3 +8,5 @@ Server helpers
 
     load_composites_plugin
     connect_to_or_start_server
+    upload_continuous_fiber_composite_files_to_server
+    upload_short_fiber_composite_files_to_server
