+++ src/ansys/dpf/composites/example_helper/__init__.py
@@ -8,7 +8,6 @@ import urllib.request
 
 import ansys.dpf.core as dpf
 
-from .._typing_helper import PATH as _PATH
 from ..data_sources import (
     CompositeDefinitionFiles,
     ContinuousFiberCompositesFiles,
@@ -16,63 +15,6 @@ from ..data_sources import (
 )
 
 EXAMPLE_REPO = "https://github.com/ansys/example-data/raw/master/pydpf-composites/"
-
-
-def upload_short_fiber_composite_files_to_server(
-    data_files: ShortFiberCompositesFiles, server: dpf.server
-) -> ShortFiberCompositesFiles:
-    """Upload short fiber composites files to server.
-
-    Parameters
-    ----------
-    data_files
-    server
-    """
-
-    def upload(filename: _PATH) -> str:
-        return cast(str, dpf.upload_file_in_tmp_folder(filename, server=server))
-
-    return ShortFiberCompositesFiles(
-        rst=upload(data_files.engineering_data),
-        dsdat=upload(data_files.dsdat),
-        engineering_data=upload(data_files.engineering_data),
-    )
-
-
-def upload_continuous_fiber_composite_files_to_server(
-    data_files: ContinuousFiberCompositesFiles, server: dpf.server
-) -> ContinuousFiberCompositesFiles:
-    """Upload continuous fiber composites files to server.
-
-    Note: If server.local_server == True the data_files are returned unmodified.
-
-    Parameters
-    ----------
-    data_files
-    server
-    """
-    if server.local_server:
-        return data_files
-
-    def upload(filename: _PATH) -> _PATH:
-        return cast(str, dpf.upload_file_in_tmp_folder(filename, server=server))
-
-    all_composite_files = {}
-    for key, composite_files_by_scope in data_files.composite.items():
-        composite_definition_files = CompositeDefinitionFiles(
-            definition=upload(composite_files_by_scope.definition),
-        )
-
-        if composite_files_by_scope.mapping is not None:
-            composite_definition_files.mapping = upload(composite_files_by_scope.mapping)
-
-        all_composite_files[key] = composite_definition_files
-
-    return ContinuousFiberCompositesFiles(
-        rst=upload(data_files.rst),
-        engineering_data=upload(data_files.engineering_data),
-        composite=all_composite_files,
-    )
 
 
 @dataclass
