+++ src/ansys/dpf/composites/server_helpers/__init__.py
@@ -6,5 +6,14 @@ the DPF Composites plugin.
 
 from ._connect_to_or_start_server import connect_to_or_start_server
 from ._load_plugin import load_composites_plugin
+from ._upload_files_to_server import (
+    upload_continuous_fiber_composite_files_to_server,
+    upload_short_fiber_composite_files_to_server,
+)
 
-__all__ = ("load_composites_plugin", "connect_to_or_start_server")
+__all__ = (
+    "load_composites_plugin",
+    "connect_to_or_start_server",
+    "upload_short_fiber_composite_files_to_server",
+    "upload_continuous_fiber_composite_files_to_server",
+)
