+++ tests/composite_model_test.py
@@ -11,7 +11,6 @@ from ansys.dpf.composites.data_sources import (
     CompositeDefinitionFiles,
     get_composite_files_from_workbench_result_folder,
 )
-from ansys.dpf.composites.example_helper import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.failure_criteria import CombinedFailureCriterion, MaxStressCriterion
 from ansys.dpf.composites.layup_info import LayerProperty, get_analysis_ply_index_to_name_map
 from ansys.dpf.composites.layup_info.material_properties import MaterialProperty
@@ -81,8 +80,6 @@ def test_basic_functionality_of_composite_model(dpf_server):
     timer = Timer()
 
     files = get_data_files()
-    files = upload_continuous_fiber_composite_files_to_server(data_files=files, server=dpf_server)
-    timer.add("After Upload files")
 
     composite_model = CompositeModel(files, server=dpf_server)
     timer.add("After Setup model")
@@ -153,9 +150,6 @@ def test_assembly_model(dpf_server):
 
     solid_label = "Setup 3_solid"
     shell_label = "Setup 4_shell"
-
-    files = upload_continuous_fiber_composite_files_to_server(data_files=files, server=dpf_server)
-    timer.add("After Upload files")
 
     composite_model = CompositeModel(files, server=dpf_server)
     timer.add("After Setup model")
@@ -297,7 +291,6 @@ def test_assembly_model(dpf_server):
 def test_failure_measures(dpf_server):
     """Verify that all failure measure names are compatible with the backend"""
     files = get_data_files()
-    files = upload_continuous_fiber_composite_files_to_server(data_files=files, server=dpf_server)
 
     composite_model = CompositeModel(files, server=dpf_server)
     combined_failure_criterion = CombinedFailureCriterion(
@@ -315,8 +308,6 @@ def test_failure_measures(dpf_server):
 def test_composite_model_element_scope(dpf_server):
     """Ensure that the element IDs of the scope can be of any type (e.g. np.int)"""
     files = get_data_files()
-
-    files = upload_continuous_fiber_composite_files_to_server(data_files=files, server=dpf_server)
 
     composite_model = CompositeModel(files, server=dpf_server)
     cfc = CombinedFailureCriterion("max stress", failure_criteria=[MaxStressCriterion()])
@@ -337,8 +328,6 @@ def test_composite_model_element_scope(dpf_server):
 def test_composite_model_named_selection_scope(dpf_server):
     """Ensure that the scoping by Named Selection is supported"""
     files = get_data_files()
-
-    files = upload_continuous_fiber_composite_files_to_server(data_files=files, server=dpf_server)
 
     composite_model = CompositeModel(files, server=dpf_server)
