+++ tests/helper.py
@@ -6,12 +6,10 @@ import ansys.dpf.core as dpf
 from ansys.dpf.core import DataSources, Field, MeshedRegion, Operator
 
 from ansys.dpf.composites.data_sources import CompositeDefinitionFiles, get_composites_data_sources
-from ansys.dpf.composites.example_helper import (
-    ContinuousFiberCompositesFiles,
-    upload_continuous_fiber_composite_files_to_server,
-)
+from ansys.dpf.composites.example_helper import ContinuousFiberCompositesFiles
 from ansys.dpf.composites.layup_info import add_layup_info_to_mesh
 from ansys.dpf.composites.layup_info.material_operators import get_material_operators
+from ansys.dpf.composites.server_helpers import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.unit_system import get_unit_system
