+++ tests/layup_properties_test.py
@@ -2,9 +2,9 @@ from ansys.dpf.core import Operator
 import pytest
 
 from ansys.dpf.composites.data_sources import get_composites_data_sources
-from ansys.dpf.composites.example_helper import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.layup_info import LayupPropertiesProvider, add_layup_info_to_mesh
 from ansys.dpf.composites.layup_info.material_operators import get_material_operators
+from ansys.dpf.composites.server_helpers import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.unit_system import get_unit_system
 
 from .helper import get_basic_shell_files
