+++ tests/performance_test.py
@@ -7,7 +7,6 @@ import pytest
 
 from ansys.dpf.composites._indexer import FieldIndexerWithDataPointer
 from ansys.dpf.composites.data_sources import CompositeDefinitionFiles, get_composites_data_sources
-from ansys.dpf.composites.example_helper import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.layup_info import (
     LayupPropertiesProvider,
     add_layup_info_to_mesh,
@@ -18,6 +17,7 @@ from ansys.dpf.composites.layup_info.material_properties import (
     MaterialProperty,
     get_constant_property_dict,
 )
+from ansys.dpf.composites.server_helpers import upload_continuous_fiber_composite_files_to_server
 from ansys.dpf.composites.unit_system import get_unit_system
 
 from .helper import ContinuousFiberCompositesFiles, Timer, setup_operators
